+++ NEWS.md
@@ -1,5 +1,9 @@
 # statsExpressions 0.2.0.9000
 
+NEW FEATURES
+
+  - New functions to carry out meta-analysis: `expr_meta_parametric`.
+
 # statsExpressions 0.2.0
 
 BREAKING CHANGES